Initial ChromiumOS commit.

+++ pwx_test_kernel_pkgs/pwx_test_kernel_pkgs_one_container.sh
@@ -14,11 +14,15 @@ Usage: pwx_test_kernel_pkgs [options] pkg_files...
         --help
 	--leave-containers-running
 	--logdir=logdir                [default: based on distribution]
+	--prepare-build
 	--pfxuse=pxfuse_src_dir        [default: download tempoary
 				        directory from github]
 	--release=dist_releaes         Which releaes of the OS distribution
                                        to use.  Mandatory.
 	--releases=dist_releases       Ignored
+	--skip-build
+	--skip-cleanup
+	--skip-load
 EOF
 }
 
@@ -36,6 +40,10 @@ make_args=""
 force=false
 logdir=$PWD
 pxfuse_dir=""
+prepare_build=false
+skip_build=true
+skip_cleanup=true
+skip_load=true
 
 exit_handler() {
     rm -rf "$local_tmp_dir"
@@ -57,6 +65,10 @@ while [[ $# -gt 0 ]] ; do
 	--pxfuse=* ) pxfuse_dir=${1#--pxfuse=} ;;
 	--release=* ) distro_release=${1#--release=} ;;
 	--releases=* ) true ;;	# Ignore
+	--prepare-build ) prepare_build=true ;;
+	--skip-build ) skip_build=true ;;
+	--skip-cleanup ) skip_cleanup=true ;;
+	--skip-load ) skip_load=true ;;
 	--* ) usage >&2 ; exit 1 ;;
 	* ) break ;;
     esac
@@ -112,6 +124,12 @@ guess_utsname_from_headers_dir() {
     local headers_dir="$1"
     local guess_utsname
 
+    guess_utsname=$(in_container cat "${headers_dir}/include/config/kernel.release")
+    if [[ $? = 0 ]] && [[ -n "$guess_utsname" ]] ; then
+	echo "${guess_utsname}"
+	return 0
+    fi
+
     guess_utsname=${headers_dir#/usr/src/}
     guess_utsname=${guess_utsname#kernels/}
     guess_utsname=${guess_utsname#linux-headers-}
@@ -125,26 +143,20 @@ guess_utsname_from_headers_dir() {
     esac
 }
 
-test_kernel_pkgs_func() {
-    local container_tmpdir result_logdir
-    local result filename real dirname basename headers_dir
-    local pkg_names deps_unfiltered dep_names arg guess_utsname
-    local dep_filtered
-    local export_dir export_pkgs_dir export_module_dir
-    local container_tmpdir=/tmp/test-portworx-kernels.$$
-    local pxfuse_dir pxd_version
-    local make_args=
+load_container() {
+}
 
-    for arg in "$@" ; do
-	echo "    $arg"
-    done >&2
-    echo "(end of arguments)" >&2
-    echo "" >&2
+
+# Usage: test_kernel_pkgs_load pxfuse_dir files...
+test_kernel_pkgs_load() {
+    local container_tmpdir
+    local result filename real dirname basename headers_dir
+    local pkg_names deps_unfiltered dep_names dep_filtered
+    local container_tmpdir=/tmp/test-portworx-kernels.$$
+    local pxfuse_dir
 
     pxfuse_dir="$1"
-    result_logdir="$2"
-    make_args="$3"
-    shift 3
+    shift
 
     in_container rm -rf "$container_tmpdir"
     in_container mkdir -p "$container_tmpdir/pxfuse_dir" "$container_tmpdir/header_pkgs"
@@ -178,8 +190,34 @@ test_kernel_pkgs_func() {
     if [[ $result != 0 ]] ; then
 	uninstall_pkgs $pkg_names
 	in_container rm -rf "$container_tmpdir"
-	return $result
     fi
+
+    return $result
+}
+
+test_kernel_pkgs_build() {
+    local container_tmpdir result_logdir
+    local result headers_dir
+    local pkg_names deps_unfiltered dep_names guess_utsname
+    local dep_filtered
+    local export_dir export_pkgs_dir export_module_dir
+    local container_tmpdir=/tmp/test-portworx-kernels.$$
+    local pxfuse_dir pxd_version
+    local make_args=
+
+    pxfuse_dir="$1"
+    result_logdir="$2"
+    make_args="$3"
+    shift 3
+
+    pkg_names=$(pkg_files_to_names "$@")
+    deps_unfiltered=$(pkg_files_to_dependencies "$@")
+
+    dep_names=""
+    for dep in $deps_unfiltered ; do
+	dep_filtered=$(filter_word "$dep" $pkg_names)
+	dep_names="$dep_names $dep_filtered"
+    done
 
     headers_dir=$(pkg_files_to_kernel_dirs "$@" | sort -u | tail -1)
     # Use "tail" to get the last kernel directory that is alphabetically
@@ -197,14 +235,23 @@ test_kernel_pkgs_func() {
         return 1
     fi
 
-    in_container sh -c \
+    result=0
+    if $prepare_build ; then 
+	${distro}_prepare_build "${container_tmpdir}" "$@"
+	result=$?
+    fi
+    if [[ $result = 0 ]] ; then
+	in_container sh -c \
                  "cd ${container_tmpdir}/pxfuse_dir && \
                   autoreconf && \
                   ./configure && \
                   make KERNELPATH=$headers_dir $make_args"
-    # make KERNELPATH=$headers_dir CC=\"gcc -fno-pie\"
 
-    result=$?
+	# make KERNELPATH=$headers_dir CC=\"gcc -fno-pie\"
+
+	result=$?
+    fi
+    
     if [[ "$result" = 0 ]] ; then
         in_container tar -C "${container_tmpdir}/pxfuse_dir" -c px.ko |
             tar -C "${result_logdir}" -xpv
@@ -225,11 +272,49 @@ test_kernel_pkgs_func() {
         cp "${result_logdir}/px.ko" "${export_module_dir}/"
 
     fi # result = 0
+
     touch "${result_logdir}/ran_test"
 
-    uninstall_pkgs $pkg_names
-    in_container rm -rf "$container_tmpdir"
-    dist_clean_up_container
+    return $result
+}
+
+test_kernel_pkgs_func() {
+    local container_tmpdir result_logdir
+    local result filename headers_dir
+    local pkg_names deps_unfiltered dep_names arg guess_utsname
+    local dep_filtered
+    local export_dir export_pkgs_dir export_module_dir
+    local container_tmpdir=/tmp/test-portworx-kernels.$$
+    local pxfuse_dir pxd_version
+    local make_args=
+
+    for arg in "$@" ; do
+	echo "    $arg"
+    done >&2
+    echo "(end of arguments)" >&2
+    echo "" >&2
+
+    pxfuse_dir="$1"
+    result_logdir="$2"
+    make_args="$3"
+    shift 3
+
+    if ! $skip_download ; then
+	test_kernel_pkgs_download "$pxfuse_dir" "$@" || return $?
+    fi
+
+    if ! $skip_build ; then
+	test_kernel_pkgs_build \
+	    "$pxfuse_dir" "$result_logdir" "$make_args" "$@"
+	result=$?
+    fi
+
+    if ! $skip_cleanup ; then
+	pkg_names=$(pkg_files_to_names "$@")
+	uninstall_pkgs $pkg_names
+	in_container rm -rf "$container_tmpdir"
+	dist_clean_up_container
+    fi
 
     echo "test_kernel_pkgs_func: build_exit_code=$result" >&2
     # if [[ "$result" != 0 ]] ; then
